+++ include/ZBE/core/io/NetMessage.h
@@ -12,7 +12,7 @@
 
 #include <cstdint>
 
-#include <SDL2/SDL_net.h>
+#include <SDL_net.h>
 
 #include "ZBE/SDL/system/SDLSocket.h"
 
@@ -33,12 +33,12 @@ public:
    *  \param status Status of the input (pressed, moved, etc.).
    *  \param time When the input was changed.
    */
-  NetMessage(Uint32 sequence, SDLSocket socket, std::vector<char> msg, uint64_t time) : sequence(sequence), socket(socket), msg(msg), time(time) {}
+  NetMessage(uint64_t sequence, SDLSocket socket, std::vector<char> msg, uint64_t time) : sequence(sequence), socket(socket), msg(msg), time(time) {}
 
   /** \brief Returns the sequence number the network message.
    *  \return The sequence number.
    */
-  Uint32 getSequence() const { return sequence;}
+  uint64_t getSequence() const { return sequence;}
 
   /** \brief Returns the socket that received the message.
    *  \return Socket that received the message.
@@ -61,7 +61,7 @@ public:
   bool operator<(const NetMessage& rhs) const {return time < rhs.getTime();}
 
 private:
-  Uint32 sequence;
+  uint64_t sequence;
   SDLSocket socket;
   std::vector<char> msg;
   uint64_t time;
